feature: only allow todo owners to update and delete their todo

+++ src/main/java/com/danielbontii/tpms/services/impl/TodoServiceImpl.java
@@ -29,6 +29,7 @@ public class TodoServiceImpl implements TodoService {
     private final UserRepository userRepository;
 
     private static final String TODO_WITH_ID_NOT_FOUND = "Todo with id %d not found";
+    private static final String ACCESS_DENIED = "Access Denied";
 
     @Override
     public List<Todo> findAll(Authentication authentication) {
@@ -48,7 +49,7 @@ public class TodoServiceImpl implements TodoService {
 
         if (!AuthorizationUtils.authorizesTodoManipulation(authentication, todo) &&
                 !AuthorizationUtils.permits(authentication, Authorities.ADMIN)) {
-            throw new AccessDeniedException("Access Denied");
+            throw new AccessDeniedException(ACCESS_DENIED);
         }
 
         return todo;
@@ -60,7 +61,7 @@ public class TodoServiceImpl implements TodoService {
         User todoOwner = userRepository.findById(todoRequest.getUserId())
                 .orElseThrow(() -> new NotFoundException("Invalid user id"));
 
-        todoRepository.findByTitleIgnoreCase(todoRequest.getTitle()).ifPresent(
+        todoRepository.findByTitleIgnoreCaseAndUserIdEquals(todoRequest.getTitle(), todoOwner.getId()).ifPresent(
                 todo -> {
                     throw new AlreadyExistsException("Todo with title " + todoRequest.getTitle() + " exists");
                 });
@@ -72,27 +73,35 @@ public class TodoServiceImpl implements TodoService {
 
     @Override
     @Transactional
-    public boolean deleteById(Long id) {
+    public boolean deleteById(Long id, Authentication authentication) {
 
         Todo todoToDelete = todoRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException(TODO_WITH_ID_NOT_FOUND.formatted(id)));
+
+        if (!AuthorizationUtils.authorizesTodoManipulation(authentication, todoToDelete)) {
+            throw new AccessDeniedException(ACCESS_DENIED);
+        }
         todoRepository.delete(todoToDelete);
 
         return true;
     }
 
     @Override
-    public Todo update(TodoUpdateInput todoUpdateInput) {
+    public Todo update(TodoUpdateInput todoUpdateInput, Authentication authentication) {
         Long todoToUpdateId = todoUpdateInput.getId();
         Todo todoToUpdate = todoRepository.findById(todoToUpdateId)
                 .orElseThrow(() -> new NotFoundException(TODO_WITH_ID_NOT_FOUND.formatted(todoToUpdateId)));
+
+        if (!AuthorizationUtils.authorizesTodoManipulation(authentication, todoToUpdate)) {
+            throw new AccessDeniedException(ACCESS_DENIED);
+        }
 
         Todo updatedTodo = TodoMapper.toUpdatedTodo(todoToUpdate, todoUpdateInput);
 
         String updatedTitle = updatedTodo.getTitle();
 
-        todoRepository.findByTitleIgnoreCase(updatedTitle).ifPresent(
-                todo -> {
+        todoRepository.findByTitleIgnoreCaseAndIdEqualsAndUserIdEquals(updatedTitle, todoToUpdate.getId(), todoToUpdate.getUser().getId())
+                .ifPresent(todo -> {
                     throw new AlreadyExistsException("Todo with title " + updatedTitle + " exists");
                 });
         return todoRepository.save(updatedTodo);
